requirements.txt
requests
python-dotenv
matplotlib
networkx
pyvis